+++ relay-event-normalization/src/normalize/span/tag_extraction.rs
@@ -1,7 +1,9 @@
 //! Logic for persisting items into `span.sentry_tags` and `span.measurements` fields.
 //! These are then used for metrics extraction.
+use std::borrow::Cow;
 use std::collections::{BTreeMap, BTreeSet};
 use std::fmt::Write;
+use std::net::IpAddr;
 use std::ops::ControlFlow;
 
 use once_cell::sync::Lazy;
@@ -15,7 +17,9 @@ use sqlparser::ast::Visit;
 use sqlparser::ast::{ObjectName, Visitor};
 use url::Url;
 
-use crate::span::description::{normalize_domain, scrub_span_description};
+use crate::span::description::{
+    concatenate_host_and_port, scrub_domain_name, scrub_span_description,
+};
 use crate::utils::{
     extract_transaction_op, http_status_code_from_span, MAIN_THREAD_NAME, MOBILE_SDKS,
 };
@@ -67,6 +71,7 @@ pub enum SpanTagKey {
     AppStartType,
     ReplayId,
     CacheHit,
+    TraceStatus,
 }
 
 impl SpanTagKey {
@@ -109,6 +114,7 @@ impl SpanTagKey {
             SpanTagKey::OsName => "os.name",
             SpanTagKey::AppStartType => "app_start_type",
             SpanTagKey::ReplayId => "replay_id",
+            SpanTagKey::TraceStatus => "trace.status",
         }
     }
 }
@@ -227,6 +233,10 @@ fn extract_shared_tags(event: &Event) -> BTreeMap<SpanTagKey, String> {
     if let Some(trace_context) = event.context::<TraceContext>() {
         if let Some(op) = extract_transaction_op(trace_context) {
             tags.insert(SpanTagKey::TransactionOp, op.to_lowercase().to_owned());
+        }
+
+        if let Some(status) = trace_context.status.value() {
+            tags.insert(SpanTagKey::TraceStatus, status.to_string());
         }
     }
 
@@ -358,16 +368,25 @@ pub fn extract_tags(
                     })
                 }) {
                     Some(domain)
-                } else if let Some(server_host) = span
+                } else if let Some(server_address) = span
                     .data
                     .value()
                     .and_then(|data| data.server_address.value())
                     .and_then(|value| value.as_str())
                 {
-                    let lowercase_host = server_host.to_lowercase();
-                    let (domain, port) = match lowercase_host.split_once(':') {
+                    let lowercase_address = server_address.to_lowercase();
+
+                    // According to OTel semantic conventions the server port should be in a separate property, called `server.port`, but incoming data sometimes disagrees
+                    let (domain, port) = match lowercase_address.split_once(':') {
                         Some((domain, port)) => (domain, port.parse::<u16>().ok()),
-                        None => (server_host, None),
+                        None => (server_address, None),
+                    };
+
+                    // Leave IP addresses alone. Scrub qualified domain names
+                    let domain = if domain.parse::<IpAddr>().is_ok() {
+                        Cow::Borrowed(domain)
+                    } else {
+                        scrub_domain_name(domain)
                     };
 
                     if let Some(url_scheme) = span
@@ -378,10 +397,11 @@ pub fn extract_tags(
                     {
                         span_tags.insert(
                             SpanTagKey::RawDomain,
-                            format!("{url_scheme}://{lowercase_host}"),
+                            format!("{url_scheme}://{lowercase_address}"),
                         );
                     }
-                    normalize_domain(domain, port)
+
+                    Some(concatenate_host_and_port(Some(domain.as_ref()), port).into_owned())
                 } else {
                     None
                 }
@@ -1452,24 +1472,22 @@ LIMIT 1
         let tags_1 = get_value!(span_1.sentry_tags).unwrap();
         let tags_2 = get_value!(span_2.sentry_tags).unwrap();
 
-        // Descriptions with loopback IPs preserve the IP and port but strip the URL path
+        // Allow loopback IPs
         assert_eq!(
             tags_1.get("description").unwrap().as_str(),
             Some("POST http://127.0.0.1:10007")
         );
-        // Domains of loopback IP descriptions preserve the IP and port
         assert_eq!(
             tags_1.get("domain").unwrap().as_str(),
             Some("127.0.0.1:10007")
         );
 
-        // Descriptions with non-loopback IPs scrub the IP naively
+        // Scrub other IPs
         assert_eq!(
             tags_2.get("description").unwrap().as_str(),
-            Some("GET http://*.8.8")
+            Some("GET http://*.*.*.*")
         );
-        // Domains of non-loopback IP descriptions are omitted
-        assert_eq!(tags_2.get("domain"), None);
+        assert_eq!(tags_2.get("domain").unwrap().as_str(), Some("*.*.*.*"));
     }
 
     #[test]
@@ -1632,6 +1650,49 @@ LIMIT 1
         assert_eq!(
             tags.get(&SpanTagKey::BrowserName),
             Some(&"Chrome".to_string())
+        );
+    }
+
+    #[test]
+    fn test_extract_trace_status() {
+        let json = r#"
+
+            {
+                "type": "transaction",
+                "platform": "python",
+                "start_timestamp": "2021-04-26T07:59:01+0100",
+                "timestamp": "2021-04-26T08:00:00+0100",
+                "transaction": "foo",
+                "contexts": {
+                    "trace": {
+                        "status": "ok"
+                    }
+                },
+                "spans": [
+                    {
+                        "op": "resource.script",
+                        "span_id": "bd429c44b67a3eb1",
+                        "start_timestamp": 1597976300.0000000,
+                        "timestamp": 1597976302.0000000,
+                        "trace_id": "ff62a8b040f340bda5d830223def1d81"
+                    }
+                ]
+            }
+        "#;
+
+        let mut event = Annotated::<Event>::from_json(json)
+            .unwrap()
+            .into_value()
+            .unwrap();
+
+        extract_span_tags_from_event(&mut event, 200);
+
+        let span = &event.spans.value().unwrap()[0];
+        let tags = span.value().unwrap().sentry_tags.value().unwrap();
+
+        assert_eq!(
+            tags.get("trace.status"),
+            Some(&Annotated::new("ok".to_string()))
         );
     }